+++ release/components/body/body.component.js
@@ -18,6 +18,7 @@ var DataTableBody = (function () {
         this.state = state;
         this.onRowClick = new core_1.EventEmitter();
         this.onRowSelect = new core_1.EventEmitter();
+        // set this before onInit for fast renders
         renderer.setElementClass(element.nativeElement, 'datatable-body', true);
     }
     Object.defineProperty(DataTableBody.prototype, "selectEnabled", {
@@ -86,9 +87,6 @@ var DataTableBody = (function () {
         this.sub.add(this.state.onSortChange.subscribe(function () {
             _this.scroller.setOffset(0);
         }));
-    };
-    DataTableBody.prototype.trackRowBy = function (index, obj) {
-        return obj.$$index;
     };
     DataTableBody.prototype.onBodyScroll = function (props) {
         this.state.offsetY = props.scrollYPos;
@@ -178,7 +176,9 @@ var DataTableBody = (function () {
         setTimeout(function () { return _this.state.options.loadingIndicator = false; }, 500);
     };
     DataTableBody.prototype.rowClicked = function (event, index, row) {
-        var clickType = event.type === 'dblclick' ? types_1.ClickType.double : types_1.ClickType.single;
+        var clickType = event.type === 'dblclick' ?
+            types_1.ClickType.double :
+            types_1.ClickType.single;
         this.onRowClick.emit({ type: clickType, event: event, row: row });
         this.selectRow(event, index, row);
     };
@@ -195,6 +195,7 @@ var DataTableBody = (function () {
         }
     };
     DataTableBody.prototype.selectRow = function (event, index, row) {
+        var _this = this;
         if (!this.selectEnabled)
             return;
         var multiShift = this.state.options.selectionType === types_1.SelectionType.multiShift;
@@ -203,14 +204,14 @@ var DataTableBody = (function () {
         if (multiShift || multiClick) {
             if (multiShift && event.shiftKey) {
                 var selected = this.state.selected.slice();
-                selections = utils_1.selectRowsBetween(selected, this.rows, index, this.prevIndex);
+                selections = utils_1.selectRowsBetween(selected, this.rows, index, this.prevIndex, function (r, s) { return _this.state.getRowSelectedIdx(r, s); });
             }
             else if (multiShift && !event.shiftKey) {
                 selections.push(row);
             }
             else {
                 var selected = this.state.selected.slice();
-                selections = utils_1.selectRows(selected, row);
+                selections = utils_1.selectRows(selected, row, function (r, s) { return _this.state.getRowSelectedIdx(r, s); });
             }
         }
         else {
@@ -220,9 +221,8 @@ var DataTableBody = (function () {
         this.onRowSelect.emit(selections);
     };
     DataTableBody.prototype.ngOnDestroy = function () {
-        if (this.sub) {
+        if (this.sub)
             this.sub.unsubscribe();
-        }
     };
     __decorate([
         core_1.Output(), 
@@ -247,7 +247,7 @@ var DataTableBody = (function () {
     DataTableBody = __decorate([
         core_1.Component({
             selector: 'datatable-body',
-            template: "\n    <div>\n      <datatable-progress\n        *ngIf=\"state.options.loadingIndicator\">\n      </datatable-progress>\n      <div\n        scroller\n        (onScroll)=\"onBodyScroll($event)\"\n        *ngIf=\"state.rows.length\"\n        [rowHeight]=\"state.options.rowHeight\"\n        [scrollbarV]=\"state.options.scrollbarV\"\n        [scrollbarH]=\"state.options.scrollbarH\"\n        [count]=\"state.rowCount\"\n        [scrollHeight]=\"state.scrollHeight\"\n        [limit]=\"state.options.limit\"\n        [scrollWidth]=\"state.columnGroupWidths.total\">\n        <datatable-row-wrapper \n          *ngFor=\"let row of rows; let i = index; trackBy: trackRowBy\"\n          [ngStyle]=\"getRowsStyles(row)\"\n          [style.height]=\"getRowHeight(row) + 'px'\"\n          [row]=\"row\">\n          <datatable-body-row\n            [attr.tabindex]=\"i\"\n            [style.height]=\"state.options.rowHeight +  'px'\"\n            (click)=\"rowClicked($event, i, row)\"\n            (dblclick)=\"rowClicked($event, i, row)\"\n            (keydown)=\"rowKeydown($event, i, row)\"\n            [row]=\"row\"\n            [class.datatable-row-even]=\"row.$$index % 2 === 0\"\n            [class.datatable-row-odd]=\"row.$$index % 2 !== 0\">\n          </datatable-body-row>\n        </datatable-row-wrapper>\n      </div>\n      <div\n        class=\"empty-row\"\n        *ngIf=\"!rows.length\"\n        [innerHTML]=\"state.options.emptyMessage\">\n      </div>\n    </div>\n  "
+            template: "\n    <div>\n      <datatable-progress\n        *ngIf=\"state.options.loadingIndicator\">\n      </datatable-progress>\n      <div\n        scroller\n        (onScroll)=\"onBodyScroll($event)\"\n        *ngIf=\"state.rows.length\"\n        [rowHeight]=\"state.options.rowHeight\"\n        [scrollbarV]=\"state.options.scrollbarV\"\n        [scrollbarH]=\"state.options.scrollbarH\"\n        [count]=\"state.rowCount\"\n        [scrollHeight]=\"state.scrollHeight\"\n        [limit]=\"state.options.limit\"\n        [scrollWidth]=\"state.columnGroupWidths.total\">\n        <datatable-row-wrapper \n          *ngFor=\"let row of rows; let i = index; trackBy: row?.$$index\"\n          [ngStyle]=\"getRowsStyles(row)\"\n          [style.height]=\"getRowHeight(row) + 'px'\"\n          [row]=\"row\">\n          <datatable-body-row\n            [attr.tabindex]=\"i\"\n            [style.height]=\"state.options.rowHeight +  'px'\"\n            (click)=\"rowClicked($event, i, row)\"\n            (dblclick)=\"rowClicked($event, i, row)\"\n            (keydown)=\"rowKeydown($event, i, row)\"\n            [row]=\"row\"\n            [class.datatable-row-even]=\"row.$$index % 2 === 0\"\n            [class.datatable-row-odd]=\"row.$$index % 2 !== 0\">\n          </datatable-body-row>\n        </datatable-row-wrapper>\n      </div>\n      <div\n        class=\"empty-row\"\n        *ngIf=\"!rows.length\"\n        [innerHTML]=\"state.options.emptyMessage\">\n      </div>\n    </div>\n  "
         }), 
         __metadata('design:paramtypes', [services_1.StateService, core_1.ElementRef, core_1.Renderer])
     ], DataTableBody);